making changes to file structures
